# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Control API",
        "description": "API para sistema de control de riesgo en operaciones de trading",
        "contact": {
            "email": "soporte@riskcontrol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/trades": {
            "get": {
                "tags": [
                    "Trades"
                ],
                "summary": "Listar trades",
                "description": "Obtener lista de trades",
                "operationId": "221ce7b87337731c6bd725bccc861beb",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "query",
                        "description": "Filtrar por ID de cuenta",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "open",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrar por tipo",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "BUY",
                                "SELL"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de trades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Trade"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trades"
                ],
                "summary": "Crear trade",
                "description": "Crear nuevo trade",
                "operationId": "21ec71c9e8993e494f7407bc39f6242a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "account_id",
                                    "type",
                                    "volume",
                                    "open_time",
                                    "status"
                                ],
                                "properties": {
                                    "account_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "BUY",
                                            "SELL"
                                        ],
                                        "example": "BUY"
                                    },
                                    "volume": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.5
                                    },
                                    "open_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "close_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "open_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.23456
                                    },
                                    "close_price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "closed"
                                        ],
                                        "example": "open"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trade creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trade created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Trade"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/trades/{id}": {
            "get": {
                "tags": [
                    "Trades"
                ],
                "summary": "Obtener trade por ID",
                "description": "Obtener trade por ID",
                "operationId": "3d23d4e4bffa20617f52388f926f61e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del trade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trade obtenido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trade"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Trades"
                ],
                "summary": "Actualizar trade",
                "description": "Actualizar trade",
                "operationId": "68e92b82c6509bd9417ca390bffd57c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del trade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "close_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "close_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.24567,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "closed"
                                        ],
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trade actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trade updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Trade"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trades"
                ],
                "summary": "Eliminar trade",
                "description": "Eliminar trade",
                "operationId": "e00f7f1608545a53cafdef9084b6d4de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del trade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trade eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trade deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/trades/{id}/close": {
            "post": {
                "tags": [
                    "Trades"
                ],
                "summary": "Cerrar trade",
                "description": "Cerrar trade",
                "operationId": "8e902cda648c58bd2b31dd4d9b886f06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del trade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "close_price"
                                ],
                                "properties": {
                                    "close_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.24567
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trade cerrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trade closed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Trade"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Respuesta Paginada",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Trade": {
                "title": "Operación de Trading",
                "description": "Representa una operación de trading realizada por un usuario",
                "required": [
                    "account_id",
                    "type",
                    "volume",
                    "open_time",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "Identificador único",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "account_id": {
                        "description": "ID de la cuenta asociada",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Tipo de operación",
                        "type": "string",
                        "enum": [
                            "BUY",
                            "SELL"
                        ],
                        "example": "BUY"
                    },
                    "volume": {
                        "description": "Volumen de la operación",
                        "type": "number",
                        "format": "float",
                        "example": 1.5
                    },
                    "open_time": {
                        "description": "Fecha y hora de apertura de la operación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "close_time": {
                        "description": "Fecha y hora de cierre de la operación",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "open_price": {
                        "description": "Precio de apertura",
                        "type": "number",
                        "format": "float",
                        "example": 1.23456
                    },
                    "close_price": {
                        "description": "Precio de cierre",
                        "type": "number",
                        "format": "float",
                        "example": 1.24567,
                        "nullable": true
                    },
                    "status": {
                        "description": "Estado de la operación",
                        "type": "string",
                        "enum": [
                            "open",
                            "closed"
                        ],
                        "example": "open"
                    },
                    "metadata": {
                        "description": "Metadatos adicionales de la operación",
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "Unauthorized": {
                "description": "No autorizado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Clave API inválida"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Recurso no encontrado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Not Found"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Recurso no encontrado"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}